Move functions to library instead of TypeScript file next to component
